requirements-full.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-gnupg==0.5.2
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
websockets==12.0
duckdb==1.0.0
scikit-learn==1.5.0
lightgbm==4.3.0
optuna==3.6.1
mlflow==2.14.2
faiss-cpu==1.8.0
openai==1.35.7
yfinance==0.2.40
APScheduler==3.10.4
redis==5.0.6
python-telegram-bot==21.4
weasyprint==61.0
jinja2==3.1.4
python-dotenv==1.0.1
orjson==3.10.6
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
ruff==0.5.0
mypy==1.10.0